Update Silero VAD model with simplified audio processing (#80)
Updates the VAD system to use the new Silero VAD v2 model, which
provides
  improved voice activity detection without requiring extensive spectral
  feature analysis..
  
- Before the old V1 was only able to achieve 0 - 34 probability range.
Resulting in alot of post processing functions that dragged the RTFx
down
- This time the conversion was more plannedout and resulted in a better
overal performance with a broader probability range from 0 to 99.
- Some side trainings to offset missing squeeze-excitation submodules
that were not available in the silero vad jit pytorch file.
- Currently supports stateless mode however statefulness would improve
the accuracy by about 10-15% ish