Adding support for phonetic and alias replacement  (#140)
### Why is this change needed?
This change adds alias replacement for text and phonetic in the text.

> This is the [Kokoro](/kˈOkəɹO/) TTS model. It supports the
[Misaki](/misˈɑki/) [G2P](G to P) engine for better currency, time and
number support. Here are some examples. The item costs $5.23. The
current time is 2:30 and the value of pi is 3.14 to 2 decimal places. It
also supports alias replacement so things like [Dr.](Doctor) sound
better and direct phonetic replacement as in, you say
[tomato](/təmˈɑːtQ/), I say [tomato](/təmˈAɾO/).

---------

Co-authored-by: Brandon Weng <18161326+BrandonWeng@users.noreply.github.com>