Padding blank audio to the back of short audio files (#234)
- Fix #232
- Mainly used 1 audio clip for testing because only 51.wav (in #232) was
able to reproduce the blank transcript
- Adding 1 second of blank audio to the end and trimming 1 second from
the beginning stops the transcription from being blank. Trimming 1
second from the beginning and adding 1 second to the end achieves the
same effect which shows that the bug isn't soley based on length (could
still be related to length).
- Adding/trimming either beginning/end also stops a blank transcript.
- Adding too little audio to the end still results in a blank
transcript. The smallest end‑padding that still produced a blank
transcript was 0.0399s. Anything larger than that stopped the blank
output, which lines up with the point where the file reaches the next
encoder frame.
- Fix: For short clips, I made it so that the end gets extra blank audio
to reach the next encoder frame.